persistence context created

+++ CarsharingSystem/Model/Booking.cs
@@ -5,7 +5,7 @@ public class Booking
     public const int PlatformFee = 5;
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public BookingStatus Status { get; set; }
 
     public Renter Renter { get; set; }
